Repairing static functions and moved polish notation
