--- group1/PolishNotation/specs.h
@@ -1,5 +1,0 @@
-#ifndef __SPECHELPERS__
-#define __SPECHELPERS__
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "doctest.h"
-#endif
